move the irc module up into the root
