--- src/irc/mod.rs
@@ -1,8 +1,0 @@
-mod message;
-mod prefix;
-
-/// Collection of types used by the IRC client
-pub mod types {
-    pub use super::message::Message;
-    pub use super::prefix::Prefix;
-}
